+++ libs/graph/resources/clj-kondo.exports/com.biffweb/graph/hooks/biff/graph.clj
@@ -3,15 +3,20 @@
 
 (defn defresolver [{:keys [node]}]
   (let [[_ sym _opts & args] (:children node)
-        fns                  (if (api/vector-node? (first args))
+        fns                  (cond
+                               (api/vector-node? (first args))
                                [(api/list-node
                                  (list* (api/token-node 'fn)
                                         args))]
+
+                               (api/keyword-node? (first args))
                                (keep-indexed
                                 (fn [i arg]
                                   (when (odd? i)
                                     arg))
-                                args))]
+                                args)
+
+                               :else args)]
     {:node (api/list-node
             [(api/token-node 'def)
              sym